python/django-rest/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.1
django-cors-headers==2.1.0
django-filter==1.1.0
django-jet==1.0.7
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
psycopg2==2.7.3.2
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
